Update CardLegality deletebehavior to ensure one-to-one relation is kept and no cascading happens when deleting a card legality entry

+++ CardAtlas.Server/Models/Data/Card.cs
@@ -79,7 +79,7 @@ public class Card
 
 	[ForeignKey("CardLegalityId")]
 	public CardLegality Legality { get; set; } = null!;
-	public required int CardLegalityId { get; set; }
+	public required long CardLegalityId { get; set; }
 
 	[MinLength(1)]
 	[MaxLength(3)]
